Tests updated and fixed Test cases extended for exclude and include swaggers
