+++ src/test/java/io/goodforgod/micronaut/swagger/api/SwaggerClient.java
@@ -1,9 +1,11 @@
-package micronaut.swagger.api;
+package io.goodforgod.micronaut.swagger.api;
+
 
 import io.micronaut.http.HttpResponse;
 import io.micronaut.http.MediaType;
 import io.micronaut.http.annotation.Get;
 import io.micronaut.http.client.annotation.Client;
+
 
 /**
  * @author Anton Kurako (GoodforGod)
