+++ src/test/java/io/goodforgod/micronaut/swagger/api/service/YamlMergedTests.java
@@ -1,20 +1,20 @@
-package micronaut.swagger.api.service;
+package io.goodforgod.micronaut.swagger.api.service;
 
-import io.micronaut.test.annotation.MicronautTest;
-import micronaut.swagger.api.model.Resource;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 
-import javax.inject.Inject;
+import io.goodforgod.micronaut.swagger.api.model.Resource;
+import io.goodforgod.micronaut.swagger.api.model.URIResource;
+import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
+import jakarta.inject.Inject;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
 
 /**
- * Description in progress
- *
  * @author Anton Kurako (GoodforGod)
  * @since 25.9.2020
  */
@@ -27,8 +27,8 @@ class YamlMergedTests extends Assertions {
     @Test
     void bothMerged() throws URISyntaxException {
         final List<Resource> resources = List.of(
-                new Resource(new URI("test-1.yml"), 20),
-                new Resource(new URI("META-INF/swagger/swagger.yml"), 10));
+                URIResource.of(new URI("mock/test-1.yml")),
+                URIResource.of(new URI("META-INF/swagger/swagger.yml")));
 
         final Map<Object, Object> merged = merger.merge(resources);
         assertNotNull(merged);
@@ -44,7 +44,7 @@ class YamlMergedTests extends Assertions {
 
     @Test
     void singleIsNotMerged() throws URISyntaxException {
-        final List<Resource> resources = List.of(new Resource(new URI("test-1.yml"), 20));
+        final List<Resource> resources = List.of(URIResource.of(new URI("mock/test-1.yml")));
 
         final Map<Object, Object> merged = merger.merge(resources);
         assertNotNull(merged);
